+++ TP/src/tp/Comun.java
@@ -4,7 +4,7 @@ public class Comun extends Transporte {
 
 	protected int limite;
 	
-	Comun(String patente, double volumen, double valor, int limite) {
+	Comun(String patente, int volumen, int valor, int limite) {
 		super(patente, volumen, valor);
 		this.limite = limite;
 	}
